refactor cluster command resp parser to adhere to new protocol restrictions
